requirements.txt
Django
djongo
faker
bcrypt
django[argon2]
pillow
django-crispy-forms
misaka
django-braces
